requirements.txt
django
psycopg2-binary
djangorestframework
dj-rest-auth
djangorestframework-simplejwt